Refactors landing page and simulator structure
Renames landing page to index.html and simulator to sim.html to improve navigation and clarity.

Adds a new preview image for the simulator on the landing page.

Includes an SEO-focused section on the landing page to describe the simulator's purpose and benefits.

Adds Umami analytics for tracking usage.

Updates the simulator's title and description for better SEO.